refactor/move process hook
