+++ src/common/hooks.ts
@@ -1,7 +1,6 @@
 import { QueryDefinition } from '@reduxjs/toolkit/dist/query';
 import { UseQueryHookResult } from '@reduxjs/toolkit/dist/query/react/buildHooks';
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { toast } from 'react-hot-toast';
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router-dom';
 import type { RootState, AppDispatch } from '../app/store';
@@ -11,7 +10,6 @@ import {
   isValidParam,
   setParams,
 } from '../features/params/paramsSlice';
-import { ProcessState } from './api/collectionsApi';
 
 declare global {
   interface Window {
@@ -61,41 +59,6 @@ export const useBackoffPolling = <
   }, [opts.baseInterval, count, pollCondition, opts.rate, result, shouldPoll]);
 
   return result;
-};
-
-export const useProcessStatePolling = <
-  StateKey extends string,
-  Result extends { [processStateKey in StateKey]: ProcessState },
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  R extends UseQueryHookResult<QueryDefinition<unknown, any, any, Result>>
->(
-  result: R,
-  processStateKeys: StateKey[],
-  options?: { baseInterval?: number; rate?: number; skipPoll?: boolean }
-) => {
-  useBackoffPolling<R>(
-    result,
-    (result) => {
-      for (let i = 0; i < processStateKeys.length; i++) {
-        const processStateKey = processStateKeys[i];
-        if (result.error || !result.data?.[processStateKey]) return false;
-        if (result.data[processStateKey] === 'processing') {
-          continue;
-        } else if (result.data[processStateKey] === 'complete') {
-          return false;
-        } else if (result.data[processStateKey] === 'failed') {
-          toast('ProcessState Polling Failed, see console for more info');
-          // eslint-disable-next-line no-console
-          console.error('ProcessState Polling Failed', { result });
-          return false;
-        } else {
-          return false;
-        }
-      }
-      return true;
-    },
-    options
-  );
 };
 
 export const ignoredParameterWarning = (ignored: string[]) =>
