+++ src/features/collections/data_products/Microtrait.tsx
@@ -15,10 +15,11 @@ import {
 import { parseError } from '../../../common/api/utils/parseError';
 import { DataViewLink } from '../../../common/components';
 import { Pagination, usePageBounds } from '../../../common/components/Table';
-import { useAppDispatch, useProcessStatePolling } from '../../../common/hooks';
+import { useAppDispatch } from '../../../common/hooks';
 import { formatNumber } from '../../../common/utils/stringUtils';
 import classes from '../Collections.module.scss';
 import { useMatchId, useGenerateSelectionId } from '../collectionsSlice';
+import { useProcessStatePolling } from '../hooks';
 import { HeatMap, HeatMapCallback, MAX_HEATMAP_PAGE } from './HeatMap';
 
 const getUPAFromEncoded = (encoded: string) => encoded.replaceAll('_', '/');
