+++ src/features/collections/hooks.ts
@@ -1,6 +1,10 @@
+import { QueryDefinition } from '@reduxjs/toolkit/dist/query';
+import { UseQueryHookResult } from '@reduxjs/toolkit/dist/query/react/buildHooks';
 import { useMemo } from 'react';
+import { toast } from 'react-hot-toast';
+import { ProcessState } from '../../common/api/collectionsApi';
 import { getNarratives } from '../../common/api/searchApi';
-import { useAppSelector } from '../../common/hooks';
+import { useAppSelector, useBackoffPolling } from '../../common/hooks';
 import {
   FilterContext,
   useFilters,
@@ -88,5 +92,40 @@ export const useTableViewParams = (
       view.selected,
       view.selection_mark,
     ]
+  );
+};
+
+export const useProcessStatePolling = <
+  StateKey extends string,
+  Result extends { [processStateKey in StateKey]: ProcessState },
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  R extends UseQueryHookResult<QueryDefinition<unknown, any, any, Result>>
+>(
+  result: R,
+  processStateKeys: StateKey[],
+  options?: { baseInterval?: number; rate?: number; skipPoll?: boolean }
+) => {
+  useBackoffPolling<R>(
+    result,
+    (result) => {
+      for (let i = 0; i < processStateKeys.length; i++) {
+        const processStateKey = processStateKeys[i];
+        if (result.error || !result.data?.[processStateKey]) return false;
+        if (result.data[processStateKey] === 'processing') {
+          continue;
+        } else if (result.data[processStateKey] === 'complete') {
+          return false;
+        } else if (result.data[processStateKey] === 'failed') {
+          toast('ProcessState Polling Failed, see console for more info');
+          // eslint-disable-next-line no-console
+          console.error('ProcessState Polling Failed', { result });
+          return false;
+        } else {
+          return false;
+        }
+      }
+      return true;
+    },
+    options
   );
 };
